Finished Market Area
-Can return
-Uses a separate tile_vec from the main screen
-Moved some of markets functionality out of main into file that used to only handle transitions.
-Moved save/load functions out of main into its own file.
